+++ src/components/utils/statustree.rs
@@ -3,7 +3,7 @@ use super::filetree::{
 };
 use anyhow::Result;
 use asyncgit::StatusItem;
-use std::{cmp, collections::BTreeSet};
+use std::{cell::Cell, cmp, collections::BTreeSet};
 
 //TODO: use new `filetreelist` crate
 
@@ -16,6 +16,8 @@ pub struct StatusTree {
 	// some folders may be folded up, this allows jumping
 	// over folders which are folded into their parent
 	pub available_selections: Vec<usize>,
+
+	pub window_height: Cell<Option<usize>>,
 }
 
 ///
@@ -27,6 +29,8 @@ pub enum MoveSelection {
 	Right,
 	Home,
 	End,
+	PageDown,
+	PageUp,
 }
 
 #[derive(Copy, Clone, Debug)]
@@ -143,6 +147,15 @@ impl StatusTree {
 				}
 				MoveSelection::Home => SelectionChange::new(0, false),
 				MoveSelection::End => self.selection_end(),
+				MoveSelection::PageUp => self.selection_page_updown(
+					selection,
+					(0..=selection).rev(),
+				),
+				MoveSelection::PageDown => self
+					.selection_page_updown(
+						selection,
+						selection..(self.tree.len()),
+					),
 			};
 
 			let changed_index =
@@ -279,6 +292,25 @@ impl StatusTree {
 			new_index = new_index.saturating_sub(1);
 			new_index = cmp::min(new_index, items_max);
 		}
+
+		SelectionChange::new(new_index, false)
+	}
+
+	fn selection_page_updown(
+		&self,
+		current_index: usize,
+		range: impl Iterator<Item = usize>,
+	) -> SelectionChange {
+		let page_size = self.window_height.get().unwrap_or(0);
+
+		let new_index = range
+			.filter(|index| {
+				self.available_selections.contains(index)
+					&& self.is_visible_index(*index)
+			})
+			.take(page_size)
+			.last()
+			.unwrap_or(current_index);
 
 		SelectionChange::new(new_index, false)
 	}